fixed home viewmodel code

+++ core/database/src/main/java/com/xero/interview/data/domain/use_case/account_record/CountAccountRecordUseCase.kt
@@ -4,7 +4,7 @@ import com.xero.interview.data.domain.repo.AccountRecordRepo
 import javax.inject.Inject
 
 open class CountAccountRecordUseCase @Inject constructor(private val repo: AccountRecordRepo) {
-    open suspend operator fun invoke(bankAccountId: Long): Int {
-        return repo.countRecords(bankAccountId)
+    open suspend operator fun invoke(bankAccountId: Int): Int {
+        return repo.countRecords(bankAccountId.toLong())
     }
 }
